Collection:  - moved Collection tests to its own file  - added tests for is() and equals()  - added docstrings for is() and equals() to clarify the difference
Series:
 - added docstrings for is() and equals() to clarify the difference
