+++ docs/series.md
@@ -118,7 +118,8 @@ series.avg("NASA_north", d => d.in);  // 250
         * [.sizeValid()](#TimeSeries+sizeValid)
         * [.count()](#TimeSeries+count) ⇒ <code>number</code>
     * _static_
-        * [.equal()](#TimeSeries.equal)
+        * [.equal(series1, series2)](#TimeSeries.equal) ⇒ <code>bool</code>
+        * [.is(series1, series2)](#TimeSeries.is) ⇒ <code>bool</code>
         * [.sum(data, seriesList, fieldSpec)](#TimeSeries.sum) ⇒ <code>[TimeSeries](#TimeSeries)</code>
 
 <a name="TimeSeries+toJSON"></a>
@@ -258,10 +259,30 @@ Returns the number of rows in the series. (Same as size())
 **Returns**: <code>number</code> - Size of the series  
 <a name="TimeSeries.equal"></a>
 
-### TimeSeries.equal()
-STATIC
+### TimeSeries.equal(series1, series2) ⇒ <code>bool</code>
+Static function to compare two TimeSeries to each other. If the TimeSeries
+are of the same instance as each other then equals will return true.
 
 **Kind**: static method of <code>[TimeSeries](#TimeSeries)</code>  
+**Returns**: <code>bool</code> - result  
+**Params**
+
+- series1 <code>[TimeSeries](#TimeSeries)</code>
+- series2 <code>[TimeSeries](#TimeSeries)</code>
+
+<a name="TimeSeries.is"></a>
+
+### TimeSeries.is(series1, series2) ⇒ <code>bool</code>
+Static function to compare two TimeSeries to each other. If the TimeSeries
+are of the same value as each other then equals will return true.
+
+**Kind**: static method of <code>[TimeSeries](#TimeSeries)</code>  
+**Returns**: <code>bool</code> - result  
+**Params**
+
+- series1 <code>[TimeSeries](#TimeSeries)</code>
+- series2 <code>[TimeSeries](#TimeSeries)</code>
+
 <a name="TimeSeries.sum"></a>
 
 ### TimeSeries.sum(data, seriesList, fieldSpec) ⇒ <code>[TimeSeries](#TimeSeries)</code>
