+++ lib/collection.js
@@ -833,11 +833,28 @@ var Collection = function (_BoundedIn) {
          * STATIC
          */
 
+        /**
+         * Static function to compare two collections to each other. If the collections
+         * are of the same instance as each other then equals will return true.
+         * @param  {Collection} collection1
+         * @param  {Collection} collection2
+         * @return {bool} result
+         */
+
     }], [{
         key: "equal",
         value: function equal(collection1, collection2) {
             return collection1._type === collection2._type && collection1._eventList === collection2._eventList;
         }
+
+        /**
+         * Static function to compare two collections to each other. If the collections
+         * are of the same value as each other then equals will return true.
+         * @param  {Collection} collection1
+         * @param  {Collection} collection2
+         * @return {bool} result
+         */
+
     }, {
         key: "is",
         value: function is(collection1, collection2) {
