+++ lib/series.js
@@ -708,11 +708,28 @@ var TimeSeries = function () {
          * STATIC
          */
 
+        /**
+         * Static function to compare two TimeSeries to each other. If the TimeSeries
+         * are of the same instance as each other then equals will return true.
+         * @param  {TimeSeries} series1
+         * @param  {TimeSeries} series2
+         * @return {bool} result
+         */
+
     }], [{
         key: "equal",
         value: function equal(series1, series2) {
             return series1._data === series2._data && series1._collection === series2._collection;
         }
+
+        /**
+         * Static function to compare two TimeSeries to each other. If the TimeSeries
+         * are of the same value as each other then equals will return true.
+         * @param  {TimeSeries} series1
+         * @param  {TimeSeries} series2
+         * @return {bool} result
+         */
+
     }, {
         key: "is",
         value: function is(series1, series2) {
