+++ src/series.js
@@ -482,11 +482,25 @@ class TimeSeries {
      * STATIC
      */
 
+     /**
+      * Static function to compare two TimeSeries to each other. If the TimeSeries
+      * are of the same instance as each other then equals will return true.
+      * @param  {TimeSeries} series1
+      * @param  {TimeSeries} series2
+      * @return {bool} result
+      */
     static equal(series1, series2) {
         return (series1._data === series2._data &&
                 series1._collection === series2._collection);
     }
 
+     /**
+      * Static function to compare two TimeSeries to each other. If the TimeSeries
+      * are of the same value as each other then equals will return true.
+      * @param  {TimeSeries} series1
+      * @param  {TimeSeries} series2
+      * @return {bool} result
+      */
     static is(series1, series2) {
         return (Immutable.is(series1._data, series2._data) &&
                 Collection.is(series1._collection, series2._collection));
